requirements.txt
# Elasticsearch and Search
elasticsearch>=8.0.0

# CrewAI and Agent Framework (Python 3.11+ version)
crewai>=0.80.0
crewai-tools>=0.12.0
langchain>=0.1.0

# ColPali for Image Processing (with compatible versions)
colpali-engine>=0.3.6
transformers>=4.46.1,<4.47.0  # Required by colpali-engine
torch>=2.7.0,<2.8.0  # Compatible with torchvision
torchvision>=0.22.0

# Image Processing
Pillow>=10.0.0

# PDF Processing
PyMuPDF>=1.23.0
pdf2image>=1.16.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
tokenizers>=0.13.2,<=0.20.3  # Required by chromadb

# Data Processing
numpy>=1.26.0,<2.0.0  # Compatible with transformers

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.25.0

# Demo App
streamlit>=1.28.0
mcp>=0.9.0 